@@ -1,12 +1,9 @@
 package com.ghy.yueplayer.fragment;
 
 import android.content.Intent;
-import android.media.MediaPlayer;
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
 import android.util.Log;
 import android.view.View;
-import android.widget.AdapterView;
 import android.widget.ListView;
 import android.widget.TextView;
 import android.widget.Toast;
@@ -17,20 +14,15 @@ import com.ghy.yueplayer.adapter.OnLineMusicListAdapter;
 import com.ghy.yueplayer.api.APIS;
 import com.ghy.yueplayer.base.BaseFragment;
 import com.ghy.yueplayer.bean.OnLineListInfo;
-import com.ghy.yueplayer.network.HttpListener;
-import com.ghy.yueplayer.network.JavaBeanRequest;
+import com.ghy.yueplayer.network.RetrofitManager;
+import com.ghy.yueplayer.network.observer.EntityObserverNoBase;
 import com.ghy.yueplayer.util.AnimUtils;
 import com.github.ybq.android.spinkit.SpinKitView;
-import com.yolanda.nohttp.RequestMethod;
-import com.yolanda.nohttp.rest.Request;
-
-import java.io.IOException;
 
 import butterknife.Bind;
 import butterknife.OnClick;
 
 /**
- * A simple {@link Fragment} subclass.
  * 榜单页面
  */
 public class TopListFragment extends BaseFragment {
@@ -42,7 +34,6 @@ public class TopListFragment extends BaseFragment {
     @Bind(R.id.lv_hot_music)
     ListView mLvHotMusic;
 
-    private MediaPlayer player;
     private OnLineMusicListAdapter mOnLineMusicListAdapter;
 
     @Override
@@ -72,7 +63,6 @@ public class TopListFragment extends BaseFragment {
      */
     private void requestListMusic() {
         String url = APIS.BASE_URL_BAI_DU_MUSIC;
-        Request<OnLineListInfo> request = new JavaBeanRequest<>(url, RequestMethod.GET, OnLineListInfo.class);
         requestParams.put("from", "android");
         requestParams.put("version", "6.0.0.3");
         requestParams.put("method", APIS.BAI_DU_METHOD_LIST);
@@ -80,67 +70,40 @@ public class TopListFragment extends BaseFragment {
         requestParams.put("type", "2");
         requestParams.put("offset", "0");
         requestParams.put("size", "100");
-        requestAPI(request, getTopListListener, null);
-    }
+        getAPiService().getOnLineListInfo(url, requestParams)
+                .compose(RetrofitManager.schedulersTransformer())
+                .compose(this.bindToLifecycle())//绑定生命周期
+                .subscribe(new EntityObserverNoBase<OnLineListInfo>(getActivity(), null) {
+                    @Override
+                    public void requestCallback(OnLineListInfo onLineListInfo, String msg, int code, boolean success) {
+                        if (success) {
 
-    /**
-     * 获取歌曲列表回调
-     */
-    private HttpListener<OnLineListInfo> getTopListListener = new HttpListener<OnLineListInfo>() {
-
-        @Override
-        public void onSucceed(final OnLineListInfo onLineListInfo) {
-            Log.i("onLineMusic", "获取榜单成功-->>");
-            mSpinKitView.setVisibility(View.GONE);
-            if (onLineListInfo == null || onLineListInfo.getSong_list() == null ||
-                    onLineListInfo.getSong_list().size() == 0) {
-                Toast.makeText(getActivity(), "获取数据为空", Toast.LENGTH_SHORT).show();
-                return;
-            }
-            mOnLineMusicListAdapter = new OnLineMusicListAdapter(getActivity(), onLineListInfo);
-            mLvHotMusic.setAdapter(mOnLineMusicListAdapter);
-            mLvHotMusic.setOnItemClickListener(new AdapterView.OnItemClickListener() {
-                @Override
-                public void onItemClick(AdapterView<?> adapterView, View view, int i, long l) {
-                    //播放
-                    String musicName = onLineListInfo.getSong_list().get(i).getTitle();
-                    String artist = onLineListInfo.getSong_list().get(i).getAuthor();
-                    String imgUrl = onLineListInfo.getSong_list().get(i).getPic_big();
-                    String songId = onLineListInfo.getSong_list().get(i).getSong_id();
-
-                    String musicPath = APIS.BASE_URL_BAI_DU_MUSIC + "?method="
-                            + APIS.BAI_DU_METHOD_PLAY + "&songid=" + songId;
-
-                    if (player != null) {
-                        player.release();
-                        player = null;
-                    }
-                    player = new MediaPlayer();
-                    try {
-                        player.reset();
-                        player.setDataSource(musicPath);
-                        player.prepareAsync();
-                        player.setOnPreparedListener(new MediaPlayer.OnPreparedListener() {
-                            @Override
-                            public void onPrepared(MediaPlayer mp) {
-                                player.start();
+                            Log.i("onLineMusic", "获取榜单成功-->>");
+                            mSpinKitView.setVisibility(View.GONE);
+                            if (onLineListInfo == null || onLineListInfo.getSong_list() == null ||
+                                    onLineListInfo.getSong_list().size() == 0) {
+                                Toast.makeText(getActivity(), "获取数据为空", Toast.LENGTH_SHORT).show();
+                                return;
                             }
-                        });
-//                        player.start();
-                    } catch (IOException e) {
-                        e.printStackTrace();
+                            mOnLineMusicListAdapter = new OnLineMusicListAdapter(getActivity(), onLineListInfo);
+                            mLvHotMusic.setAdapter(mOnLineMusicListAdapter);
+                            mLvHotMusic.setOnItemClickListener((adapterView, view, i, l) -> {
+                                //播放
+                                String musicName = onLineListInfo.getSong_list().get(i).getTitle();
+                                String artist = onLineListInfo.getSong_list().get(i).getAuthor();
+                                String imgUrl = onLineListInfo.getSong_list().get(i).getPic_big();
+                                String songId = onLineListInfo.getSong_list().get(i).getSong_id();
+                                String musicPath = APIS.BASE_URL_BAI_DU_MUSIC + "?method="
+                                        + APIS.BAI_DU_METHOD_PLAY + "&songid=" + songId;
+                            });
+                        } else {
+                            Log.i("onLineMusic", "获取榜单出错-->>" + msg);
+                            mSpinKitView.setVisibility(View.GONE);
+                            Toast.makeText(getActivity(), "请求出错", Toast.LENGTH_SHORT).show();
+                        }
                     }
-                }
-            });
-        }
-
-        @Override
-        public void onFailed(int errorCode, String msg) {
-            Log.i("onLineMusic", "获取榜单出错-->>" + msg);
-            mSpinKitView.setVisibility(View.GONE);
-            Toast.makeText(getActivity(), "请求出错", Toast.LENGTH_SHORT).show();
-        }
-    };
+                });
+    }
 
     @OnClick({R.id.tv_see_all})
     public void topListClick(View view) {
@@ -154,9 +117,6 @@ public class TopListFragment extends BaseFragment {
     @Override
     public void onDestroy() {
         super.onDestroy();
-        if (player != null) {
-            player.release();
-            player = null;
-        }
     }
+
 }
